feat: modify code to  support frontend need,  make standalone task management
